requirements.txt
asn1crypto==0.24.0
astroid==2.2.4
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
coveralls==1.6.0
cryptography==2.4.2
dj-database-url==0.5.0
Django==2.1.7
django-rest-swagger==2.2.0
djangorestframework==3.9.2
docopt==0.6.2
drf-yasg==1.14.0
flasgger==0.9.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Extended==3.17.0
gunicorn==19.9.0
idna==2.8
import-string==0.1.0
inflection==0.3.1
isort==4.3.12
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
more-itertools==6.0.0
openapi-codec==1.3.2
pluggy==0.9.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.7
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
pytest==4.3.0
pytest-cov==2.6.1
pytest-django==3.4.8
#python-coveralls==2.9.1
python-dotenv==0.10.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
ruamel.yaml==0.15.89
simplejson==3.16.0
six==1.12.0
travis==0.0.3
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.1.0
Werkzeug==0.14.1
whitenoise==4.1.2
wrapt==1.11.1
